Modified Task 1 #0

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
 	try:
-		ans = int(numerator) / int(denominator)
+		ans = float(numerator) / float(denominator)
 	except ValueError:
 		print("Error: Please enter numeric values only.")
 		exit()
